hardware/ld2450_test/requirements.txt
matplotlib>=3.7
pyserial>=3.5
